stdin.h in helpers and details in EJ.1.c

+++ 02_fixed-point/Resolution example/Ej1.c
@@ -15,7 +15,7 @@ void main(void)
 	
 	c = a + b;	
 	// d = a * b;
-	d = (int32_t)(a * b);
+	d = (a * b);
 	
 	s1 = (-8) >> 2;
 	s2 = (-1) >> 5;
